+++ hdiv-config/src/main/java/org/hdiv/config/annotation/builders/SecurityConfigBuilder.java
@@ -114,6 +114,16 @@ public class SecurityConfigBuilder {
 		return this;
 	}
 
+	public SecurityConfigBuilder pentestingActive(final boolean penTestingActive) {
+		config.setPentestingActive(penTestingActive);
+		return this;
+	}
+
+	public SecurityConfigBuilder multipartIntegration(final boolean multipartIntegration) {
+		config.setMultipartIntegration(multipartIntegration);
+		return this;
+	}
+
 	public SessionExpiredConfigure sessionExpired() {
 		return sessionExpiredConfigure;
 	}