+++ hdiv-core/src/main/java/org/hdiv/filter/ValidatorFilter.java
@@ -186,7 +186,7 @@ public class ValidatorFilter extends OncePerRequestFilter {
 			boolean legal = false;
 			boolean isMultipartException = false;
 
-			if (isMultipartContent(request)) {
+			if (isMultipartContent(request) && hdivConfig.isMultipartIntegration()) {
 
 				requestWrapper.setMultipart(true);